+++ DAVEX_center/src/main/java/DavexCenter/module/file/service/FileService.java
@@ -49,7 +49,7 @@ public class FileService {
     @Autowired
     private DownloadTaskMapper downloadTaskMapper;
 
-    public Body<String> saveFile(MultipartFile file, File fileInfo, Integer applicationId,
+    public Body<String> saveFile(MultipartFile file, File fileInfo, Long applicationId,
             String base, java.sql.Timestamp expiredTime) {
 
         // 校验sha256
@@ -86,7 +86,7 @@ public class FileService {
         newOutput.setHash(fileHash);
         newOutput.setFileId(fileInfo.getUid());
         newOutput.setAgentId(fileInfo.getAgentId());
-        newOutput.setApplicationId(applicationId.longValue());
+        newOutput.setApplicationId(applicationId);
         outputMapper.insert(newOutput);
         String filePath = newOutput.getPath();
 
@@ -102,7 +102,7 @@ public class FileService {
                 fileInfo.getUid(), fileInfo.getAgentId(), fileName));
     }
 
-    public Body<List<String>> saveFiles(List<MultipartFile> files, List<File> fileInfos, Integer applicationId,
+    public Body<List<String>> saveFiles(List<MultipartFile> files, List<File> fileInfos, Long applicationId,
             String base, List<java.sql.Timestamp> expiredTimes) {
         List<String> results = new ArrayList<>();
 
@@ -152,7 +152,7 @@ public class FileService {
             newOutput.setHash(fileHash);
             newOutput.setFileId(fileInfo.getUid());
             newOutput.setAgentId(fileInfo.getAgentId());
-            newOutput.setApplicationId(applicationId.longValue());
+            newOutput.setApplicationId(applicationId);
             outputMapper.insert(newOutput);
             String filePath = newOutput.getPath();
 
@@ -174,7 +174,7 @@ public class FileService {
         return Body.success(results, "文件保存处理完成");
     }
 
-    public Body<String> fetchFileByHttp(Integer outputId, Integer applicationId, HttpServletResponse response) {
+    public Body<String> fetchFileByHttp(Long outputId, Long applicationId, HttpServletResponse response) {
 
         // 根据结果id查找结果表
         LambdaQueryWrapper<Output> queryWrapper = Wrappers.<Output>lambdaQuery()
@@ -225,7 +225,7 @@ public class FileService {
         return Body.success(String.format("获取成功，结果id: %d，文件名: %s", outputId, fileName));
     }
 
-    public Body<String> fetchFile(Integer outputId, Integer applicationId, String downloadPath) {
+    public Body<String> fetchFile(Long outputId, Long applicationId, String downloadPath) {
 
         // 根据结果id查找结果表
         LambdaQueryWrapper<Output> queryWrapper = Wrappers.<Output>lambdaQuery()
@@ -244,7 +244,7 @@ public class FileService {
 
         // 添加下载任务记录到任务表
         DownloadTask newDownloadTask = new DownloadTask();
-        newDownloadTask.setApplicationId(Long.valueOf(applicationId));
+        newDownloadTask.setApplicationId(applicationId);
         newDownloadTask.setOutputId(queryOutput.getUid());
         newDownloadTask.setDownloadTime(Timestamp.valueOf(LocalDateTime.now()));
         newDownloadTask.setType("common");
@@ -262,7 +262,7 @@ public class FileService {
         return Body.success(String.format("获取成功，结果id: %d，文件名: %s", outputId, fileName));
     }
 
-    public Body<List<Output>> queryFile(Integer applicationId) {
+    public Body<List<Output>> queryFile(Long applicationId) {
 
         LambdaQueryWrapper<Output> queryWrapper = Wrappers.<Output>lambdaQuery()
                 .eq(Output::getApplicationId, applicationId);
@@ -271,7 +271,7 @@ public class FileService {
         return Body.success(outputs, String.format("查询成功，共查询到%d个文件", fileNum));
     }
 
-    public Body<List<Output>> queryFileByIds(Integer applicationId, List<Integer> outputIds) {
+    public Body<List<Output>> queryFileByIds(Long applicationId, List<Long> outputIds) {
 
         LambdaQueryWrapper<Output> queryWrapper = Wrappers.<Output>lambdaQuery()
                 .eq(Output::getApplicationId, applicationId)
@@ -282,7 +282,7 @@ public class FileService {
         return Body.success(outputs, String.format("查询成功，共查询到%d个文件", fileNum));
     }
 
-    public Body<String> deleteFile(Integer applicationId, Integer outputId) {
+    public Body<String> deleteFile(Long applicationId, Long outputId) {
 
         // 根据文件id查找结果表
         LambdaQueryWrapper<Output> queryWrapper = Wrappers.<Output>lambdaQuery()
